// src/types/product.types.ts
export type Discount = {
  amount: number;
  percentage: number;
};

export type Product = {
  id: number;
  title: string;
  srcUrl: string;
  gallery?: string[];
  price: number;
  discount: Discount;
  rating: number;
  category: string;
  colors: string[];
  sizes: string[];
  style: string;
};